backend/requirements.txt
agno>=0.8.0
fastapi>=0.95
uvicorn[standard]>=0.22
httpx>=0.24
python-dotenv>=1.0
pydantic>=2.0
chromadb>=0.4
rich>=12.6
async-timeout>=4.0

# Testing
pytest>=7.4
pytest-asyncio>=0.21

# Optional / recommended
psycopg2-binary>=2.9  # if using Postgres for Agno storage
